cmm: Split crates
