+++ hacking/cargo-manifest-management/crates/manage-cargo-manifests/Cargo.toml
@@ -1,0 +1,23 @@
+#
+# Copyright 2023, Colias Group, LLC
+#
+# SPDX-License-Identifier: BSD-2-Clause
+#
+#
+
+[package]
+name = "manage-cargo-manifests"
+version = "0.1.0"
+authors = ["Nick Spinale <nick.spinale@coliasgroup.com>"]
+edition = "2021"
+license = "BSD-2-Clause"
+
+[dependencies]
+clap = { version = "4.4.6", features = [ "derive" ] }
+serde = { version = "1.0", features = [ "derive" ] }
+serde_json = "1.0"
+toml = "0.8.6"
+toml_edit = "0.20.4"
+similar = "2.3"
+toml-path-regex = { path = "../toml-path-regex" }
+toml-normalize = { path = "../toml-normalize" }
