+++ hacking/cargo-manifest-management/crates/manage-cargo-manifests/src/diff.rs
