+++ hacking/cargo-manifest-management/crates/toml-normalize/src/lib.rs
@@ -1,0 +1,7 @@
+mod cargo_manifest_policy;
+mod easy_policy;
+mod format;
+
+pub use cargo_manifest_policy::cargo_manifest_policy;
+pub use easy_policy::{EasyPolicy, TableRule, TableRuleOrdering};
+pub use format::{Formatter, Policy};
