+++ hacking/cargo-manifest-management/crates/toml-path-regex/src/path_segment_predicate.rs
